Refactor code to introduce ChessGame class and improve structure
The main refactor was to move game control logic from main.cpp to a new class called ChessGame. This makes the code cleaner and easier to manage. Along with this, also renamed the function 'generateMove' to 'buildMove' for clearer naming, removed commented-out code, and moved 'player_type' enum to a separate header file 'player.h'. The 'Move' structure now also includes an equality operator and a new method 'contains' is added to 'PseudoLegalMoves'.